Intial Queue

+++ app/src/main/java/deep/ryd/rydplayer/DatabasHelper.java
@@ -52,6 +52,9 @@ public class DatabasHelper extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
+        if (oldVersion==1){
+            return;
+        }
         db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
         onCreate(db);
     }
